move ParameterCoercer
This class used from HyperSchema only
