+++ lib/committee/schema_validator/hyper_schema/parameter_coercer.rb
@@ -1,6 +1,5 @@
-# TODO: Support OpenAPI3
 module Committee
-  class ParameterCoercer
+  class SchemaValidator::HyperSchema::ParameterCoercer
     def initialize(params, schema, options = {})
       @params = params
       @schema = schema
